requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
beautifulsoup4==4.7.1
blessed==1.15.0
boto3==1.9.4
botocore==1.12.4
cached-property==1.5.1
cement==2.8.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3.1
Cython==0.29.2
defusedxml==0.5.0
Django==2.0.5
django-background-tasks==1.2.0
django-bulk-update==2.2.0
django-ckeditor==5.7.1
django-cors-headers==2.4.0
django-countries==5.3.3
django-environ==0.4.5
django-extensions==2.1.2
django-filebrowser==3.11.1
django-filter==2.1.0
django-grappelli==2.12.1
django-hosts==3.0
django-graphql-jwt==0.1.13
django-import-export==1.1.0
django-imagekit==4.0.2 
django-mysql==2.4.1
django-redis==4.9.0
django-rest-knox==3.1.5
django-storages==1.7.1
django-warrant==0.1.0
django-webpack-loader==0.6.0
djangorestframework-jwt==1.11.0
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13
envs==1.2.6
future==0.16.0
graphene==2.1.3
graphene-django==2.2.0
graphene-file-upload==1.2.2
graphql-core==2.1
graphql-relay==0.4.5
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
Markdown==2.6.11
mysqlclient==1.3.13
numpy==1.15.1
oauthlib==2.1.0
pathspec==0.5.5
pilkit==2.0
Pillow==5.2.0
promise==2.2.1
pycparser==2.19
pycryptodome==3.3.1
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-jose-cryptodome==1.3.2
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.0.0
Rx==1.6.1
s3transfer==0.1.13
semantic-version==2.5.0
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
tqdm==4.26.0
urllib3==1.22
warrant==0.6.1
wcwidth==0.1.7
websocket-client==0.53.0
